requirements.txt
click==7.1.2
colorama==0.4.4
python-dotenv==0.15.0
shellingham==1.3.2
typer==0.3.2
